feat(Toolbar): show GPS heading and clarify Chinese GPS labels
Expose yaw from GPS_RAW_INT when available, rename COG to track angle in zh_CN, and translate HDOP/VDOP.

+++ src/Vehicle/FactGroups/VehicleGPSFactGroup.cc
@@ -33,7 +33,7 @@ VehicleGPSFactGroup::VehicleGPSFactGroup(QObject *parent)
     _hdopFact.setRawValue(std::numeric_limits<float>::quiet_NaN());
     _vdopFact.setRawValue(std::numeric_limits<float>::quiet_NaN());
     _courseOverGroundFact.setRawValue(std::numeric_limits<float>::quiet_NaN());
-    _yawFact.setRawValue(std::numeric_limits<int16_t>::quiet_NaN());
+    _yawFact.setRawValue(std::numeric_limits<float>::quiet_NaN());
     _spoofingStateFact.setRawValue(255);
     _jammingStateFact.setRawValue(255);
     _authenticationStateFact.setRawValue(255);
